git-svn-id: http://pubsvn.inavare.net/openscada/modules/atlantis@187 1b50e8ee-deed-0310-b1de-ab401e1c4d63

+++ da/client/test/src/org/openscada/da/client/test/config/HiveConnectionInformation.java
@@ -10,7 +10,7 @@ public class HiveConnectionInformation implements Serializable
     private static final long serialVersionUID = 4226848319802862860L;
     
     private String _host;
-    private short _port;
+    private int _port;
     
     public HiveConnectionInformation ()
     {
@@ -26,11 +26,11 @@ public class HiveConnectionInformation implements Serializable
     {
         _host = host;
     }
-    public short getPort ()
+    public int getPort ()
     {
         return _port;
     }
-    public void setPort ( short port )
+    public void setPort ( int port )
     {
         _port = port;
     }
